+++ NEWS.md
@@ -1,3 +1,5 @@
+# ReboundTools 0.1.12 (2021-03-10)
+
 * Removed `sensitivity_graphs()` for now.
 * New constants for sensitivity graphs, `sens_graph_params`.
 * Name changes: